+++ web/src/pages/InfoPage/InfoPage.tsx
@@ -1,20 +1,20 @@
 import { Metadata } from '@redwoodjs/web'
 
-import ConfigurationCell from 'src/components/ConfigurationCell'
+import InfoCell from 'src/components/InfoCell'
 
-const ConfigurationPage = () => {
+const InfoPage = () => {
   return (
     <>
-      <Metadata title="Configuration" description="Configuration page" />
+      <Metadata title="Info" description="Info page" />
       <h3 className="text-tremor-title font-bold text-tremor-content-strong dark:text-dark-tremor-content-strong">
-        Configuration
+        Info
       </h3>
       <p className="mt-2 text-tremor-default leading-6 text-tremor-content dark:text-dark-tremor-content">
-        Studio configuration settings.
+        Various studio configuration information and settings.
       </p>
-      <ConfigurationCell />
+      <InfoCell />
     </>
   )
 }
 
-export default ConfigurationPage
+export default InfoPage
